feat(icons): apply stroke-width plugin to comfy icons too
Extract comfy icon set loader into shared module, allow icon-s*
utilities to accept the comfy-- prefix, and replace any stroke-width
attribute (not just 2) so comfy icons with native widths like 1.3
or 1.95 also respond to the override.

Amp-Thread-ID: https://ampcode.com/threads/T-019e70e1-11c3-73f9-981a-5c1d6392a2c9
