+++ packages/design-system/src/css/iconifyDynamicPlugin.js
@@ -1,35 +1,12 @@
-import { readdirSync, readFileSync } from 'node:fs'
-import { basename, join, resolve } from 'node:path'
-
 import { getDynamicCSSRules } from '@iconify/tailwind4/lib/plugins/dynamic.js'
 import plugin from 'tailwindcss/plugin'
 
-const COMFY_ICONS_DIR = resolve(import.meta.dirname, '../icons')
+import { COMFY_ICON_PREFIX, loadComfyIconSet } from './comfyIconSet.js'
+
 const SCALE = 1.2
 
-function loadComfyIconSet() {
-  const icons = {}
-  for (const file of readdirSync(COMFY_ICONS_DIR)) {
-    if (!file.endsWith('.svg')) continue
-    const name = basename(file, '.svg')
-    const svg = readFileSync(join(COMFY_ICONS_DIR, file), 'utf8')
-    const svgMatch = svg.match(/<svg\b([^>]*)>([\s\S]*?)<\/svg>/i)
-    if (!svgMatch) continue
-    const viewBox = svgMatch[1].match(/\bviewBox=(['"])(.*?)\1/i)?.[2]
-    const viewBoxMatch = viewBox?.match(
-      /^(?:-?\d*\.?\d+)\s+(?:-?\d*\.?\d+)\s+(?<width>\d*\.?\d+)\s+(?<height>\d*\.?\d+)$/
-    )
-    icons[name] = {
-      body: svgMatch[2],
-      width: Number(viewBoxMatch?.groups?.width ?? '16'),
-      height: Number(viewBoxMatch?.groups?.height ?? '16')
-    }
-  }
-  return { prefix: 'comfy', icons }
-}
-
 const options = {
-  iconSets: { comfy: loadComfyIconSet() },
+  iconSets: { [COMFY_ICON_PREFIX]: loadComfyIconSet() },
   scale: SCALE
 }
