+++ packages/design-system/src/css/lucideStrokePlugin.js
@@ -1,11 +1,13 @@
-import plugin from 'tailwindcss/plugin'
 import { getIconsCSSData } from '@iconify/utils/lib/css/icons'
-import { loadIconSet } from '@iconify/tailwind4/lib/helpers/loader.js'
 import { matchIconName } from '@iconify/utils/lib/icon/name'
+import { loadIconSet } from '@iconify/tailwind4/lib/helpers/loader.js'
+import plugin from 'tailwindcss/plugin'
+
+import { COMFY_ICON_PREFIX, loadComfyIconSet } from './comfyIconSet.js'
 
 /**
- * Tailwind 4 plugin that provides lucide icon variants with configurable
- * stroke-width via class prefix.
+ * Tailwind 4 plugin that provides icon variants with configurable
+ * stroke-width via class prefix. Supports lucide and comfy icon sets.
  *
  * Usage in CSS:
  *   @plugin "./lucideStrokePlugin.js";
@@ -13,18 +15,26 @@ import { matchIconName } from '@iconify/utils/lib/icon/name'
  * Usage in templates:
  *   <i class="icon-s1-[lucide--settings]" />     <!-- stroke-width: 1 -->
  *   <i class="icon-s1.5-[lucide--settings]" />   <!-- stroke-width: 1.5 -->
- *   <i class="icon-s2.5-[lucide--settings]" />   <!-- stroke-width: 2.5 -->
+ *   <i class="icon-s2.5-[comfy--workflow]" />    <!-- stroke-width: 2.5 -->
  *
- * The default class remains stroke-width: 2.
+ * The plain `icon-[...]` class keeps each icon's native stroke-width.
  */
 
 const STROKE_WIDTHS = ['1', '1.3', '1.5', '2', '2.5']
 
-const SUPPORTED_PREFIX = 'lucide'
+const LUCIDE_PREFIX = 'lucide'
+const SUPPORTED_PREFIXES = new Set([LUCIDE_PREFIX, COMFY_ICON_PREFIX])
 
 const SCALE = 1.2
 
+const STROKE_WIDTH_ATTR_RE = /stroke-width="[^"]*"/g
+
 class InvalidIconProbeError extends Error {}
+
+function resolveIconSet(prefix) {
+  if (prefix === COMFY_ICON_PREFIX) return loadComfyIconSet()
+  return loadIconSet(prefix)
+}
 
 function getDynamicCSSRulesWithStroke(icon, strokeWidth) {
   const nameParts = icon.split(/--|:/)
@@ -32,13 +42,13 @@ function getDynamicCSSRulesWithStroke(icon, strokeWidth) {
     throw new InvalidIconProbeError(`Invalid icon name: "${icon}"`)
   }
   const [prefix, name] = nameParts
-  if (prefix !== SUPPORTED_PREFIX) {
+  if (!SUPPORTED_PREFIXES.has(prefix)) {
     throw new InvalidIconProbeError(`Unsupported icon prefix: "${prefix}"`)
   }
   if (!(prefix.match(matchIconName) && name.match(matchIconName))) {
     throw new InvalidIconProbeError(`Invalid icon name: "${icon}"`)
   }
-  const iconSet = loadIconSet(prefix)
+  const iconSet = resolveIconSet(prefix)
   if (!iconSet) {
     throw new Error(
       `Cannot load icon set for "${prefix}". Install "@iconify-json/${prefix}" as dev dependency?`
@@ -47,7 +57,7 @@ function getDynamicCSSRulesWithStroke(icon, strokeWidth) {
   const generated = getIconsCSSData(iconSet, [name], {
     iconSelector: '.icon',
     customise: (content) =>
-      content.replaceAll('stroke-width="2"', `stroke-width="${strokeWidth}"`)
+      content.replace(STROKE_WIDTH_ATTR_RE, `stroke-width="${strokeWidth}"`)
   })
   if (generated.css.length !== 1) {
     throw new Error(`Cannot find "${icon}". Bad icon name?`)
